Fri, Mar 8, 2024, 1:42 PM -05:00

+++ chapter3/filter-more/helpers.c
@@ -7,7 +7,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < height; j++)
         {
-            int average = ((image[j][i].rgbtRed + image[j][i].rgbtGreen + image[j][i].rgbtBlue) / 3);
+            float average = ((image[j][i].rgbtRed + image[j][i].rgbtGreen + image[j][i].rgbtBlue) / 3);
             image[j][i].rgbtRed = average;
             image[j][i].rgbtGreen = average;
             image[j][i].rgbtBlue = average;
